Improve concurrency
By default, Keycloak admin service uses one connection, which results in errors
if more than one thread are concurrently requesting access. In this commit, the
amount of connections used by keycloak is defaulted to the number of worker threads
for client connections, which should be somewhat of reflection of the amount of
concurrent calls that could take place.