Work in progress: Updating DWR to 3.0.2
This is an incomplete change!

The DWR libary was updated to version 3.0.2.

The Monitoring en Kraken plugins contain an DWRServlet extension, which fails to compile with the new DWR jar. I've replaced the subclass in the Monitoring plugin, which makes it compile. Sadly, the Monitoring plugin now fails when being loaded, with exceptions logged. I've not even began work in the Kraken plugin, where a similar fix is needed.